correção

+++ parcelApp/src/com/parcelinc/parcelapp/db/PagamentoDataBase.java
@@ -158,7 +158,7 @@ public class PagamentoDataBase implements DataBase<Pagamento> {
 		double valor = c.getDouble(4);
 
 		Despesa despesa = null;
-		if (idDespesa.longValue() != 0) {
+		if (idDespesa != 0) {
 			despesa = despesaDataBase.get(idDespesa);
 		}
 		Usuario usuario = null;
